requirements_mac.txt
absl-py==0.1.10
appnope==0.1.0
backcall==0.1.0
bleach==1.5.0
certifi==2018.1.18
chardet==3.0.4
decorator==4.3.0
dominate==2.3.1
entrypoints==0.2.3
futures==3.1.1
html5lib==0.9999999
idna==2.6
ipykernel==4.8.2
ipython==6.4.0
ipython-genutils==0.2.0
ipywidgets==7.2.1
jedi==0.12.0
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
Markdown==2.6.11
MarkupSafe==1.0
mistune==0.8.3
nbconvert==5.3.1
nbformat==4.4.0
notebook==5.5.0
numpy==1.14.0
pandocfilters==1.4.2
parso==0.2.1
pexpect==4.5.0
pickleshare==0.7.4
Pillow==5.0.0
prompt-toolkit==1.0.15
protobuf==3.5.1
ptyprocess==0.5.2
Pygments==2.2.0
python-dateutil==2.7.3
PyYAML==3.12
pyzmq==17.0.0
qtconsole==4.3.1
requests==2.18.4
scipy==1.0.0
Send2Trash==1.5.0
simplegeneric==0.8.1
six==1.11.0
tensorboardX==1.0
tensorflow==1.5.0
tensorflow-tensorboard==1.5.0
terminado==0.8.1
testpath==0.3.1
torch==0.4.0
torchfile==0.1.0
torchvision==0.2.1
tornado==4.5.3
traitlets==4.3.2
urllib3==1.22
visdom==0.1.8.3
wcwidth==0.1.7
websocket-client==0.47.0
Werkzeug==0.14.1
widgetsnbextension==3.2.1
